+++ src/vs/workbench/workbench.desktop.main.ts
@@ -65,6 +65,7 @@ import 'vs/workbench/services/update/electron-browser/updateService';
 import 'vs/workbench/services/issue/electron-browser/issueService';
 import 'vs/workbench/services/menubar/electron-browser/menubarService';
 import 'vs/workbench/services/extensionResourceLoader/electron-browser/extensionResourceLoaderService';
+import 'vs/workbench/services/title/electron-browser/titleService';
 
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { ICredentialsService } from 'vs/platform/credentials/common/credentials';
